Renderer: Rework render statistics monitoring
This kills a bug that made render time estimation calculations overflow.
It also makes the render time estimation a bit inaccurate. Oh well, can't win 'em all.

+++ src/utils/logging.c
@@ -67,7 +67,7 @@ void printSmartTime(unsigned long long ms) {
 
 // Print to buf a logically formatted string representing time given in milliseconds.
 // Example: "980ms", "2s", "1m 02s", "3h 12m", etc.
-void smartTime(unsigned long long milliseconds, char *buf) {
+void smartTime(uint64_t milliseconds, char *buf) {
 	time_t secs = milliseconds / 1000;
 	time_t mins = secs / 60;
 	time_t hours = (secs / 60) / 60;
